Change pk of wordform links to composite primary key
The primary key of wordform links is now a composite key. This means
that if we want to add a wordform link source, we only need to know the
wordform ids of the wordforms in question, instead of having to look up
the id after it was added. This saves a lot of overhead.