Fix update_anahashes
There was a bug in getting the wordform mapping, resulting in having
zeros for all wordform_ids. This bug was fixed.

Also the (skipped) test didn't work because of the aforementioned bug.
And to make the test work, we had to call dbsession.commit() after
updating the anahash_ids of the wordforms.

The test in question is still skipped, because we still don't have a way
of installing ticcl on travis. It can be run locally, though.