Database migration for removing the morphological paradigms table
We found out the strings with morphological information for a word are not unique,
so they can't be used as (composite) primary key for the morphological paradigms
table. Because we are changing primary keys of a table and are not interested in
keeping the data, it is more convenient to just drop the old table and re-create
it with the desired properties.